BUG: Use q->mrmlScene() in favor of this->scene
the use of this-scene in GetSubjectHierarchyNode has already been the
source of at least two segfaults. I did a search for instances of this
pattern and attempted to preemptively fix these by adding the scene as a
method parameter. This led to an explosion in methods with this new
parameter, so I switched to this q pointer based approach.