BUG: Remove qMRMLPlannerModelHierarchyModel and ButtonItemDelegate
SubjectHierarchyTreeView is not designed to operate with the setModel
method, even though it inherits this method from QTreeView. This means
we have to ditch qMRMLPlannerModelHierarchyModel.

ButtonItemDelegate was used only to handle buttons in columns with the
qMRMLPlannerModelHierarchyModel.

Some methods were moved from qMRMLPlannerModelHierarchyModel.

In order to make the cut and bend functionality available again, it will
be necessary to add buttons to the UI that trigger these actions.